Move cityDetailsActions to separate file
